WEBAPP-375: Use PageDetail in pages and rename CategoryTimeStamp
